REFACTOR Reorganize domain/product package structure
- Create subpackages by responsibility
  - vo/: Value Objects (ProductScope, PricingType, ProductPriceBreakdown)
  - pricing/: Pricing strategy (PricingStrategy)
  - availability/: Availability checking (ProductAvailabilityService, Checkers)
  - relationship/: Relationship management (RoomAllowedProduct)

- Update all import statements across codebase
- Update test files with correct imports
- Product Aggregate remains at top level

Benefits:
- Clear separation of concerns by package
- Easier navigation and file discovery
- Better cohesion within each package
- Improved onboarding for new developers

Closes #126
