+++ springProject/src/main/java/com/teambind/springproject/application/dto/request/RegisterProductRequest.java
@@ -1,7 +1,7 @@
 package com.teambind.springproject.application.dto.request;
 
 import com.teambind.springproject.domain.product.Product;
-import com.teambind.springproject.domain.product.ProductScope;
+import com.teambind.springproject.domain.product.vo.ProductScope;
 import com.teambind.springproject.domain.shared.PlaceId;
 import com.teambind.springproject.domain.shared.ProductId;
 import com.teambind.springproject.domain.shared.RoomId;
