+++ springProject/src/main/java/com/teambind/springproject/application/port/in/GetProductUseCase.java
@@ -1,7 +1,7 @@
 package com.teambind.springproject.application.port.in;
 
 import com.teambind.springproject.application.dto.response.ProductResponse;
-import com.teambind.springproject.domain.product.ProductScope;
+import com.teambind.springproject.domain.product.vo.ProductScope;
 import com.teambind.springproject.domain.shared.PlaceId;
 import com.teambind.springproject.domain.shared.ProductId;
 import com.teambind.springproject.domain.shared.RoomId;
