+++ springProject/src/main/java/com/teambind/springproject/application/service/product/GetProductService.java
@@ -4,7 +4,7 @@ import com.teambind.springproject.application.dto.response.ProductResponse;
 import com.teambind.springproject.application.port.in.GetProductUseCase;
 import com.teambind.springproject.application.port.out.ProductRepository;
 import com.teambind.springproject.domain.product.Product;
-import com.teambind.springproject.domain.product.ProductScope;
+import com.teambind.springproject.domain.product.vo.ProductScope;
 import com.teambind.springproject.domain.shared.PlaceId;
 import com.teambind.springproject.domain.shared.ProductId;
 import com.teambind.springproject.domain.shared.RoomId;
