+++ springProject/src/main/java/com/teambind/springproject/domain/product/availability/ScopedAvailabilityChecker.java
@@ -1,5 +1,6 @@
-package com.teambind.springproject.domain.product;
+package com.teambind.springproject.domain.product.availability;
 
+import com.teambind.springproject.domain.product.Product;
 import com.teambind.springproject.domain.reservationpricing.ReservationPricing;
 import java.time.LocalDateTime;
 import java.util.List;
