+++ springProject/src/test/java/com/teambind/springproject/adapter/in/web/product/ProductControllerTest.java
@@ -22,8 +22,8 @@ import com.teambind.springproject.application.port.in.GetProductUseCase;
 import com.teambind.springproject.application.port.in.QueryProductAvailabilityUseCase;
 import com.teambind.springproject.application.port.in.RegisterProductUseCase;
 import com.teambind.springproject.application.port.in.UpdateProductUseCase;
-import com.teambind.springproject.domain.product.ProductScope;
-import com.teambind.springproject.domain.product.PricingType;
+import com.teambind.springproject.domain.product.vo.ProductScope;
+import com.teambind.springproject.domain.product.vo.PricingType;
 import com.teambind.springproject.domain.shared.PlaceId;
 import com.teambind.springproject.domain.shared.ProductId;
 import com.teambind.springproject.domain.shared.RoomId;
