+++ springProject/src/test/java/com/teambind/springproject/domain/product/ProductAvailabilityServiceTest.java
@@ -8,6 +8,10 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import com.teambind.springproject.application.port.out.ReservationPricingRepository;
+import com.teambind.springproject.domain.product.availability.ProductAvailabilityService;
+import com.teambind.springproject.domain.product.pricing.PricingStrategy;
+import com.teambind.springproject.domain.product.vo.PricingType;
+import com.teambind.springproject.domain.product.vo.ProductPriceBreakdown;
 import com.teambind.springproject.domain.reservationpricing.ReservationPricing;
 import com.teambind.springproject.domain.reservationpricing.TimeSlotPriceBreakdown;
 import com.teambind.springproject.domain.shared.Money;
