+++ springProject/src/test/java/com/teambind/springproject/domain/product/ProductPriceBreakdownTest.java
@@ -1,5 +1,8 @@
 package com.teambind.springproject.domain.product;
 
+import com.teambind.springproject.domain.product.vo.ProductPriceBreakdown;
+import com.teambind.springproject.domain.product.vo.PricingType;
+
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
