+++ springProject/src/test/java/com/teambind/springproject/domain/product/ProductScopeValidationTest.java
@@ -1,5 +1,8 @@
 package com.teambind.springproject.domain.product;
 
+import com.teambind.springproject.domain.product.pricing.PricingStrategy;
+import com.teambind.springproject.domain.product.vo.ProductScope;
+
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
