+++ springProject/src/test/java/com/teambind/springproject/e2e/InventoryManagementE2ETest.java
@@ -10,8 +10,8 @@ import com.teambind.springproject.application.port.out.PricingPolicyRepository;
 import com.teambind.springproject.application.port.out.ProductRepository;
 import com.teambind.springproject.domain.pricingpolicy.PricingPolicy;
 import com.teambind.springproject.domain.product.Product;
-import com.teambind.springproject.domain.product.ProductScope;
-import com.teambind.springproject.domain.product.PricingStrategy;
+import com.teambind.springproject.domain.product.vo.ProductScope;
+import com.teambind.springproject.domain.product.pricing.PricingStrategy;
 import com.teambind.springproject.domain.shared.Money;
 import com.teambind.springproject.domain.shared.PlaceId;
 import com.teambind.springproject.domain.shared.ProductId;
