+++ springProject/src/test/java/com/teambind/springproject/e2e/RoomAllowedProductE2ETest.java
@@ -8,7 +8,7 @@ import com.teambind.springproject.application.port.out.ProductRepository;
 import com.teambind.springproject.application.port.out.RoomAllowedProductRepository;
 import com.teambind.springproject.domain.pricingpolicy.PricingPolicy;
 import com.teambind.springproject.domain.product.Product;
-import com.teambind.springproject.domain.product.PricingStrategy;
+import com.teambind.springproject.domain.product.pricing.PricingStrategy;
 import com.teambind.springproject.domain.shared.Money;
 import com.teambind.springproject.domain.shared.PlaceId;
 import com.teambind.springproject.domain.shared.ProductId;
